requirements.txt
Django==1.5.2
South==0.8.2
amqp==1.0.13
anyjson==0.3.3
argparse==1.2.1
billiard==2.7.3.31
celery==3.0.21
django-celery==3.0.21
django-sitegate==0.5.0
hiredis==0.1.1
kombu==2.5.12
python-dateutil==2.1
pytz==2013b
redis==2.7.6
six==1.3.0
wsgiref==0.1.2
